switch event number to double

+++ DataFormats/src/PATFinalStateEvent.cc
@@ -467,9 +467,9 @@ const edm::EventID& PATFinalStateEvent::evtId() const {
   return evtID_;
 }
 
-int PATFinalStateEvent::event() const {
+double PATFinalStateEvent::event() const {
   ULong64_t eventNum = evtId().event();
-  int doub = 0;
+  double doub = 0;
   return doub + eventNum;
 }
